Flask Socket Io
This is my first time using Flask Socketio but i must say that it was really simple to use...

In this simple progamme, when a button is clicked , it emits a yes , no or maybe to the brower.

This is a fairly simple progamme to understand how flask_socketio operates.